make the latest resource as shared section
