+++ src/sections/shared/resources.tsx
@@ -5,7 +5,7 @@ import { ICONS } from "@/utils/icons";
 import { Button, ButtonGroup } from "@nextui-org/react";
 import { useState } from "react";
 
-const Resources = () => {
+const LatestResources = () => {
   const [resources, setResources] = useState(resourcesData.slice(0, 4));
   const [activeTab, setActiveTab] = useState(resourcesCategories[0]);
 
@@ -41,7 +41,11 @@ const Resources = () => {
   };
   return (
     <section className="px-5 md:px-10 lg:px-16 xl:px-24 2xl:px-36 3xl:px-44 pb-10 lg:pb-14 xl:pb-20 bg-ice-blue  ">
-      <div className="flex flex-col md:flex-row gap-5 items-center justify-end mb-10">
+      <div className="flex flex-col md:flex-row gap-5 items-center justify-between mb-10">
+        <h2 className="text-2xl md:text-4xl lg:text-5xl xl:text-6xl text-[#272727] font-medium">
+          Latest resources
+        </h2>
+
         <ButtonGroup
           radius="full"
           variant="bordered"
@@ -95,4 +99,4 @@ const Resources = () => {
   );
 };
 
-export default Resources;
+export default LatestResources;
